tools/requirements.txt
Flask==1.1.1
PyStanfordDependencies==0.3.1
bllipparser==2016.9.11
ufal.udpipe==1.2.0.2
beautifulsoup4==4.7.1